+++ 42_Projects/42_Formation/libft/ft_memchr.c
@@ -6,13 +6,13 @@
 /*   By: andrefil <andrefil@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/13 16:50:46 by andrefil          #+#    #+#             */
-/*   Updated: 2023/10/26 00:25:23 by andrefil         ###   ########.fr       */
+/*   Updated: 2023/10/28 21:49:44 by andrefil         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memchr(const void *str, int ch, int num)
+void	*ft_memchr(const void *str, int ch, size_t num)
 {
 	unsigned char	*s;
 	unsigned char	c;
